Refactor FilterPanel and related components for improved organization
